+++ packages/terra-date-time-picker/src/DateTimeUtils.js
@@ -126,9 +126,7 @@ class DateTimeUtils {
    * @return {boolean} - True if the time is valid.
    */
   static isValidTime(time, hasSeconds) {
-    const timeFormat = hasSeconds ? 'HH:mm:ss' : 'HH:mm';
-    const timeMoment = moment(time, timeFormat, true);
-    return timeMoment.isValid();
+    return TimeUtil.isValidTime(time, hasSeconds);
   }
 
   /**