+++ packages/terra-time-input/src/TimeInput.jsx
@@ -133,6 +133,7 @@ class TimeInput extends React.Component {
     this.handleMeridiemInputFocus = this.handleMeridiemInputFocus.bind(this);
     this.handleMeridiemSelectFocus = this.handleMeridiemSelectFocus.bind(this);
     this.handleMeridiemButtonChange = this.handleMeridiemButtonChange.bind(this);
+    this.handleMissingHourChange = this.handleMissingHourChange.bind(this);
 
     let hour = TimeUtil.splitHour(value);
     let meridiem;
@@ -174,16 +175,23 @@ class TimeInput extends React.Component {
     };
   }
 
-  componentDidUpdate(prevProps) {
-    if (
-      this.props.value === prevProps.value
-      && this.props.variant === prevProps.variant
+  componentDidUpdate(prevProps, prevState) {
+    let skipUpdate = true;
+    const prevStateTimeValue = (this.props.showSeconds) ? `${prevState.hour}:${prevState.minute}:${prevState.second}` : `${prevState.hour}:${prevState.minute}`;
+    // Comparing partial time values on each update will result in incorrect comparisons of prevState and propsValue.
+    // Hence validating propsValue and prevState values to skip the comparison between them until time value is completely formed.
+    if (TimeUtil.isValidTime(this.props.value, this.props.showSeconds) && TimeUtil.isValidTime(prevStateTimeValue, this.props.showSeconds)) {
+      skipUpdate = this.handleMissingHourChange(prevState);
+    }
+
+    if (this.props.value === prevProps.value
+      && this.props.variant === prevProps.variant && skipUpdate
     ) {
       return;
     }
 
-    let hour = TimeUtil.splitHour(this.props.value);
     let { meridiem } = this.state;
+    let hour = TimeUtil.splitHour(this.props.value);
 
     if (this.props.variant === TimeUtil.FORMAT_12_HOUR) {
       if (!this.props.intl.messages['Terra.timeInput.am'] || !this.props.intl.messages['Terra.timeInput.pm']) {
@@ -214,6 +222,20 @@ class TimeInput extends React.Component {
       second: TimeUtil.splitSecond(this.props.value),
       meridiem,
     });
+  }
+
+  handleMissingHourChange(prevState) {
+    let hour = TimeUtil.splitHour(this.props.value);
+    if (this.props.variant === TimeUtil.FORMAT_12_HOUR) {
+      hour = TimeUtil.parseTwelveHourTime(hour, this.anteMeridiem, this.postMeridiem).hourString;
+    }
+    const minute = TimeUtil.splitMinute(this.props.value);
+    const second = TimeUtil.splitSecond(this.props.value);
+
+    if (hour !== prevState.hour || minute !== prevState.minute || second !== prevState.second) {
+      return false;
+    }
+    return true;
   }
 
   handleFocus(event) {