+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Fixed time-input componentDidUpdate for updating time-input on re-entering the same missing hour in time-input.
 
 4.9.0 - (September 6, 2019)
 ------------------